refactor(config): extract scope profile parser
