+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 - Exposed recall introspection through the typed `Orchestrator.recallIntrospection`
   coordinator and removed the duplicate intent, QMD, and graph-recall forwarding
   methods from the oversized composition root.
+- Moved scope-profile and team configuration parsing into a dedicated
+  namespace config module while preserving `parseConfig` validation and
+  defaults.
 
 ## [v9.3.652] — 2026-06-30
